fix(wp-0075): close regression clusters LISS-0204..0207
Restore anticommutator braces and algebra Call parsing; align OOP/SI/residual
suites with ADR 0116, shipped units, and linear HARD_CODES.

+++ tests/test_function_signatures_red.py
@@ -60,13 +60,13 @@ class Box {
         this.x = value
     }
 
-    fn doubled() -> State<Int> {
+    fn doubled() -> State<Float> {
         return dirac(this.x + this.x)
     }
 }
 pub fn main() -> Unit {
     Box box = Box(3)
-    State<Int> result = box.doubled()
+    State<Float> result = box.doubled()
     measure result
 }
 """
